Fix: Bad file descriptor (-1) on close.
This bug has never been reported as having caused any problems as
the result of the close(2) is ignored.

The fix is a kludge, but the bug will disappear when the code is
restructured in MOC 2.6.  For now, we need to mark it as having
been found and fixed anyway.

git-svn-id: svn://svn.daper.net/moc/branches/2.5@2821 910807d9-36e0-0310-a014-e9ea483e2ba4