Fix: Delayed reporting of "TERMINAL TOO SMALL".
It seems that performing a temporary suspend (via endwin() /
refresh()) is the accepted way to force Curses to refresh its
terminal context, and it isn't until after that that the new COLS
and LINES settings take effect.  It feels uncomfortable to have
that update happen as a side-effect of an otherwise redundant
suspend sequence, however, it appears to be the only portable way
of doing it.

Multiple refresh()s may introduce some visible screen flicker (which
the code elsewhere in MOC tries to avoid), but it has not been
observed in testing and is probably masked by the resizing operation.

Thanks to: Joan Bruguera Mico <joanbrugueram@gmail.com>
Resolves: node/2177 (MOC prints "TERMINAL TOO SMALL" when it's not)

git-svn-id: svn://svn.daper.net/moc/branches/2.5@2993 910807d9-36e0-0310-a014-e9ea483e2ba4